Added similar methods to shared files.
